+++ src/main/java/cn/taketoday/web/servlet/DispatcherServlet.java
@@ -167,7 +167,7 @@ public class DispatcherServlet implements Servlet, InitializingBean, WebApplicat
 			request.setAttribute(Constant.KEY_REQUEST_URI, requestURI);
 			requestMapping = handlerMappingRegistry.get(index);
 			// get intercepter s
-			final Integer[] interceptors = requestMapping.getInterceptors();
+			final int[] interceptors = requestMapping.getInterceptors();
 			// invoke intercepter
 			final HandlerInterceptorRegistry handlerInterceptorRegistry = getHandlerInterceptorRegistry();
 			for (Integer interceptor : interceptors) {